# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Democratizing data access through plain English.",
        "title": "Poligono API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Poligono Support",
            "url": "https://www.swagger.io/support",
            "email": "dev@poligono.xyz"
        },
        "license": {
            "name": "GNU Affero General Public License version 3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        }
    },
    "paths": {
...
        "/prompts/generate": {
            "post": {
                "description": "Create an SQL query based on a natural language prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Generate SQL Query",
                "parameters": [
                    {
                        "description": "Prompt Object",
                        "name": "Prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GenerateQueryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GenerateQueryActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/prompts/improve": {
            "post": {
                "description": "Improve an SQL query based on a natural language prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Improve SQL Query",
                "parameters": [
                    {
                        "description": "Prompt Object",
                        "name": "Prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ImproveQueryBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImproveQueryActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.GenerateQueryActivity": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "data_source_name": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                }
            }
        },
        "models.GenerateQueryBody": {
            "type": "object",
            "properties": {
                "data_source_name": {
                    "type": "string"
                },
                "execute": {
                    "type": "boolean"
                },
                "prompt": {
                    "type": "string"
                }
            }
        },
        "models.HTTPError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.ImproveQueryActivity": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "data_source_name": {
                    "type": "string"
                },
                "improved_query": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "original_query": {
                    "type": "string"
                }
            }
        },
        "models.ImproveQueryBody": {
            "type": "object",
            "properties": {
                "data_source_name": {
                    "type": "string"
                },
                "execute": {
                    "type": "boolean"
                },
                "query": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}